// apps/console/src/pages/organizations/complianceRegistries/__generated__/ComplianceRegistriesPageFragment.graphql.ts
export type ComplianceRegistryStatus = "CLOSED" | "IN_PROGRESS" | "OPEN";
import { FragmentRefs } from "relay-runtime";
export type ComplianceRegistriesPageFragment$data = {
  readonly complianceRegistries: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly actionsToBeImplemented: string | null | undefined;
        readonly area: string | null | undefined;
        readonly audit: {
          readonly framework: {
            readonly id: string;
            readonly name: string;
          };
          readonly id: string;
          readonly name: string | null | undefined;
        };
        readonly createdAt: any;
        readonly dueDate: any | null | undefined;
        readonly id: string;
        readonly lastReviewDate: any | null | undefined;
        readonly owner: {
          readonly fullName: string;
          readonly id: string;
        };
        readonly referenceId: string;
        readonly regulator: string | null | undefined;
        readonly requirement: string | null | undefined;
        readonly source: string | null | undefined;
        readonly status: ComplianceRegistryStatus;
        readonly updatedAt: any;
      };
    }>;
    readonly pageInfo: {
      readonly endCursor: any | null | undefined;
      readonly hasNextPage: boolean;
    };
    readonly totalCount: number;
  };
  readonly id: string;
  readonly " $fragmentType": "ComplianceRegistriesPageFragment";
};
export type ComplianceRegistriesPageFragment$key = {
  readonly " $data"?: ComplianceRegistriesPageFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"ComplianceRegistriesPageFragment">;
};